Update Remove-WindowsUserProfile.ps1
Implementing What-if feature

+++ Remove-WindowsUserProfile.ps1
@@ -1,5 +1,5 @@
 function Remove-WindowsProfile {
-    [CmdletBinding()]
+    [CmdletBinding(SupportsShouldProcess=$true)]
     param (
         [Parameter(Mandatory=$true,Position=0)]
         [string]$Username
@@ -8,9 +8,13 @@ function Remove-WindowsProfile {
     $SID = (Get-WmiObject -Class Win32_UserProfile | Where-Object { $_.LocalPath.split('\')[-1] -eq $Username }).SID
     if ($SID) {
         $UserPath = Join-Path -Path "C:\Users" -ChildPath $Username
-        Remove-Item -Path $UserPath -Recurse -Force
+        if ($PSCmdlet.ShouldProcess($UserPath, "Delete")) {
+            Remove-Item -Path $UserPath -Recurse -Force
+        }
         $RegPath = Join-Path -Path $RegKey -ChildPath $SID
-        Remove-Item -Path $RegPath -Recurse -Force
+        if ($PSCmdlet.ShouldProcess($RegPath, "Delete")) {
+            Remove-Item -Path $RegPath -Recurse -Force
+        }
         Write-Output "Profile for $Username has been deleted."
     } else {
         Write-Output "Profile for $Username was not found."
